+++ sphinx/source/changelog.rst
@@ -3,6 +3,7 @@ Changelog
 
 9.2b1
 -----
+* Added support to :ref:`tap2sna.py` for TZX block type 0x15 (direct recording)
 
 9.1 (2024-02-03)
 ----------------